change int to long

+++ src/Category.java
@@ -2,7 +2,7 @@ import java.util.List;
 
 public class Category {
     private String name;
-    private Integer categoruId;
+    private Long categoruId;
     private List<Item> items;
