+++ src/Item.java
@@ -3,7 +3,7 @@ public class Item {
     private String itemName;
     private String venderName;
     private Integer price; // may be BigDecimal
-    private Integer itemsId;
+    private Long itemsId;
